refactor: project structure - extract hooks into separate files - extract hook tests into separate files
